Remove addons command
Encourage users to use kubectl instead
